+++ crates/stackable-certs/src/keys/rsa.rs
@@ -13,7 +13,7 @@ const KEY_SIZE: usize = 4096;
 
 pub type Result<T, E = Error> = std::result::Result<T, E>;
 
-#[derive(Debug, Snafu)]
+#[derive(Debug, PartialEq, Snafu)]
 pub enum Error {
     #[snafu(display("failed to create RSA key"))]
     CreateKey { source: rsa::Error },